feat: move logout button to navigation
