+++ frontend/src/components/Header/Header.scss
@@ -110,6 +110,8 @@
     transition: 0.25s;
     font-size: 0.9rem;
     cursor: pointer;
+    background: none;
+    border: none;
     &.active {
       color: $secondary;
     }
@@ -118,6 +120,20 @@
     }
     @include tablet {
       font-size: 1rem;
+    }
+  }
+
+  &__button {
+    color: $white;
+    border: none;
+    @include flex-center;
+    background: none;
+    font-size: 1rem;
+    font-weight: 600;
+    text-transform: uppercase;
+    cursor: pointer;
+    &:hover {
+      color: $primary;
     }
   }
